Eliminate alternate names for \Irr
This is done to match the notational convention set in the previous
proposition better. The previous proposition had some references to \cI
for \Irr(\cC_{20}) and some direct references, so I decided to remove
the additional \mathcal letters, preferring brevity of setup to
nicer-looking subscripts on sums. But it could go the other way.