Add wrong factors of d to action stuff
The factors of d written now make all the results true only up to
scalar, and follow from not including any scalars in the definition of
action of 1 on the right. We can make sure these are correct, and then
easily see what scalars need to appear in the action.